add message response checking
Add cols to store the response and when it was received.
Add class to do the checking...

A lot of refactoring should really be done here. The DB access is
getting messy and should be swept up into more of an activerecord
pattern.

Might create a PofWebdriver class and include modules into it to do the
various site operations

Need to figure out a way to test the webdriver shit