improve Profile.messagable and supporting code
Previously this would retrieve 200 profiles that hadn't been messaged.
Now it retrieves exactly the number of profiles that we want to message.
This fixes the error in Messager where we retrieve 200 profiles, but
none of them are actually messagable.

Update the profile factory in the specs. Modify a basic profile page
when creating the profile - so that it naturally parses to have the
given interests, pof_key etc. This allows us to remove some hacks in the
specs and is generally superior.