update the way we model messages to include our own profile as sender
Add sender/recipient profile IDs to messages.

Sender will default to our own profile (see below). This means we are
now free to cache messages that come to us unsolicited - previously we
assume any message is a response to one of ours.

Add our own profile_id to the config. This will be used for all messages
that are sent to/from us.