split Topic up into Interest and TemplateMessage
Interest contains the matching data and represents the general concept.

TemplateMessage is a message that a user can send when an Interest is
matched against a profile.

This way the matching logic for interests can be shared across all
users - they just have to have a TemplateMessage for the Interest in
question.

I'm not happy with the Profile class having #interests as well
as #pof_interests. This is confusing and will be refactored later for
sure, but just commit now that things are working.