refactor Profile interests and derived fields - much less yuck now
Previously we stored the profile page html and parsed it every time we
want things like the persons Bio, Name or Interests.

Now we store these things as fields in the DB and have the ability to
reprocess the page and reset said fields as needed.

Rework Profile#interests to not totally suck!

Previously a Profile reparsed the profile page html every time it was
loaded to build up a list of matching Interest entities... it also
stored a collection of 'pof_interests' strings that it parsed off the
profile page.

Now a Profile has_many ProfileInterests which relate to Interests. When
we call Profile#parse_page_contents! it will reprocess the profile page
contents and create the ProfileInterests according to the current
Interest entities in the DB.

This simplifies things and means profile interests are stored in the db
in a usable, efficient way... amen.