add user sign up and basic message setup wizard
This is a pretty mammoth commit. The system is very much a walking
skeleton at the moment, but the first request spec is passing:

> A user registers and goes through the template message wizard

Use sorcery for the user accounts system.

Use simple_form.

Use bootstrap. Right now there is no theme and it is just used for
layout purposes - and very minimally so.